chore: allow latest Node LTS
Currently this logic prevents this package from being installed on Node 18, which is now the current LTS. This change will allow end users to install this on Node 18 and run the scripts. Alternatively, this package should consider using `>=` instead of `||` so that it can just specify a minimum node version instead of being so perscriptive.